requirements.txt
# requests
pyYAML
dvc
wget